Rename databases.ts api file to connections.ts
